% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{bmap_get_cached_address_data}
\alias{bmap_get_cached_address_data}
\title{Get Cached Address Data}
\usage{
bmap_get_cached_address_data()
}
\value{
data frame
}
\description{
Return all cached address data, as a tidy data frame.
}
\examples{
\dontrun{
df <- bmap_get_cached_address_data()
}
}
