% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{bmap_rate_limit_info}
\alias{bmap_rate_limit_info}
\title{Get Rate Limit Details}
\usage{
bmap_rate_limit_info()
}
\value{
list of length four.
}
\description{
Function that will return rate limit details related to the current 
registered API key. Values are returned as a list.
}
\details{
Function returns a list containing the following info
\itemize{
\item Current registered API key
\item Daily query limit of the current key
\item Number of daily queries remaining for the current 24 hour period
\item Date-time in which the daily query limit will reset.
}
}
\examples{
bmap_rate_limit_info()

}
