% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{bmap_remaining_daily_queries}
\alias{bmap_remaining_daily_queries}
\title{Remaining Daily Queries}
\usage{
bmap_remaining_daily_queries()
}
\value{
integer, number of queries
}
\description{
Get the number of daily queries left in the current 24 hour period, with 
the API key that is currently registered.
}
\examples{
bmap_remaining_daily_queries()

}
