% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{bmap_set_daily_rate_limit}
\alias{bmap_set_daily_rate_limit}
\title{Set a daily rate limit for an API key.}
\usage{
bmap_set_daily_rate_limit(num_limit)
}
\arguments{
\item{num_limit}{numeric value, number of daily API queries associated with 
the valid key currently being used.}
}
\value{
Function does not return a value.
}
\description{
This function will save the value passed to arg "num_limit" to the pkg 
environment value "bmap_daily_rate_limit", which is meant to be the number 
of daily queries allowed for the API key currently being used (default 
value for "bmap_daily_rate_limit" upon pkg load is 5950). Prior to using 
this function, a valid API key must be registered in the current R session, 
using function \code{\link{bmap_set_key}}.
}
\examples{
\dontrun{
bmap_set_daily_rate_limit(20000)
}

}
