% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{bmap_set_key}
\alias{bmap_set_key}
\title{Set Baidu Map API Key}
\usage{
bmap_set_key(key)
}
\arguments{
\item{key}{char string, valid Baidu Maps API key.}
}
\value{
Function does not return a value.
}
\description{
Function for setting a Baidu Map API key. Key will be used in all API calls.
}
\examples{
bmap_set_key("some_valid_key_str")

}
